+++ radioactivedecay/__init__.py
@@ -5,6 +5,6 @@ decay data from ICRP Publication 107. It solves the decay chain differential
 equations analytically using NumPy and SciPy linear algebra routines.
 """
 
-__version__ = "0.0.7"
+__version__ = "0.0.8"
 
 from radioactivedecay.decayfunctions import *